whisper/grpc_service/requirements.txt
openai-whisper
grpcio==1.60.0
grpcio-tools==1.60.0
nats-py==2.6.0
